requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
boto3==1.42.54
botocore==1.42.54
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
multidict==6.7.1
propcache==0.4.1
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.46
twilio==9.10.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
yarl==1.22.0
